+++ components/AzTable.vue
@@ -22,7 +22,7 @@
           @click="emits('addToggle')"
           icon="tabler:plus"
         >
-          {{ t('add') }}
+          {{ t("add") }}
         </AzButton>
         <slot name="addForm" />
       </Tooltip>
@@ -32,12 +32,23 @@
       </Tooltip>
       <Divider type="vertical" />
       <Tooltip title="ریست جدول" size="small">
-        <AzButton type="link" size="small" @click="resetTable" icon="tabler:refresh" />
+        <AzButton
+          type="link"
+          size="small"
+          @click="resetTable"
+          icon="tabler:refresh"
+        />
       </Tooltip>
     </template>
 
     <template
-      #customFilterDropdown="{ setSelectedKeys, selectedKeys, confirm, clearFilters, column }"
+      #customFilterDropdown="{
+        setSelectedKeys,
+        selectedKeys,
+        confirm,
+        clearFilters,
+        column,
+      }"
     >
       <div class="p-4">
         <RangePicker
@@ -53,7 +64,9 @@
           :placeholder="`جستجو در ${t(column.dataIndex)}`"
           :value="selectedKeys[0]"
           class="mb-4"
-          @change="(e) => setSelectedKeys(e.target.value ? [e.target.value] : [])"
+          @change="
+            (e) => setSelectedKeys(e.target.value ? [e.target.value] : [])
+          "
         />
         <div class="grid grid-cols-2 gap-2">
           <AzButton
@@ -61,9 +74,11 @@
             @click="() => handleSubmit(selectedKeys, confirm, column.dataIndex)"
             icon="tabler:search"
           >
-            {{ t('filter') }}
+            {{ t("filter") }}
           </AzButton>
-          <AzButton @click="() => handleReset(clearFilters)">{{ t('reset') }}</AzButton>
+          <AzButton @click="() => handleReset(clearFilters)">{{
+            t("reset")
+          }}</AzButton>
         </div>
       </div>
     </template>
@@ -98,41 +113,48 @@
 </template>
 
 <script setup lang="ts">
-import type { ColumnsType, TablePaginationConfig, TableProps } from 'ant-design-vue/es/table'
+import type {
+  ColumnsType,
+  TablePaginationConfig,
+  TableProps,
+} from "ant-design-vue/es/table";
 import type {
   FilterValue,
   SorterResult,
-  TableCurrentDataSource
-} from 'ant-design-vue/lib/table/interface'
-import { Table, Divider, Input, RangePicker, Tooltip } from 'ant-design-vue/es'
-import { useFullscreen } from '@vueuse/core'
-import { ref, onMounted, reactive, computed, watch } from 'vue'
-import { formatDate } from '@/core/utils'
-import { AzButton, AzFullScreen, AzStatus } from '@/core/components'
-import { Icon } from '@iconify/vue/dist/iconify.js'
-import { useI18n } from 'vue-i18n'
-import { isFunction, isString, truncate } from 'lodash'
-import dayjs from 'dayjs'
-import type { PaginationConfig } from 'ant-design-vue/es/pagination'
-import { SortDirEnum } from '@/core/enums'
+  TableCurrentDataSource,
+} from "ant-design-vue/lib/table/interface";
+import { Table, Divider, Input, RangePicker, Tooltip } from "ant-design-vue/es";
+import { useFullscreen } from "@vueuse/core";
+import { ref, onMounted, reactive, computed, watch } from "vue";
+import { formatDate } from "@/core/utils";
+import { AzButton, AzFullScreen, AzStatus } from "@/core/components";
+import { Icon } from "@iconify/vue/dist/iconify.js";
+import { useI18n } from "vue-i18n";
+import { isFunction, isString, truncate } from "lodash";
+import dayjs from "dayjs";
+import type { PaginationConfig } from "ant-design-vue/es/pagination";
+import { SortDirEnum } from "@/core/enums";
 
 interface Props {
-  columns: ColumnsType
-  fetch: Function
-  reloadFlag: boolean
+  columns: ColumnsType;
+  fetch: Function;
+  reloadFlag: boolean;
 }
 
-const azTableRef = ref<HTMLElement | null>(null)
-const { toggle: toggleFullScreen, isFullscreen } = useFullscreen(azTableRef)
-const props = defineProps<Props>()
-const emits = defineEmits(['addToggle'])
-const data = ref([])
-const loading = ref(false)
-const { t } = useI18n()
+const azTableRef = ref<HTMLElement | null>(null);
+const { toggle: toggleFullScreen, isFullscreen } = useFullscreen(azTableRef);
+const props = defineProps<Props>();
+const emits = defineEmits(["addToggle"]);
+const data = ref([]);
+const loading = ref(false);
+const { t } = useI18n();
 
 const computedColumns = computed(() => {
-  return [{ title: t('row'), dataIndex: 'row', key: 'row', width: 'max-content' }, ...props.columns]
-})
+  return [
+    { title: t("row"), dataIndex: "row", key: "row", width: "max-content" },
+    ...props.columns,
+  ];
+});
 
 const pagination = reactive({
   showSizeChanger: true,
@@ -141,150 +163,148 @@ const pagination = reactive({
   pageSizeOptions: [5, 10, 20, 30, 50],
   current: 1,
   pageSize: 5,
-  total: 0
-})
+  total: 0,
+});
 
 const state = reactive<any>({
   value: undefined,
-  label: undefined
-})
-const currentData = ref({})
-const fetchData = async () => {
-  loading.value = true
+  label: undefined,
+});
+const currentData = ref({});
+const fetchData = async (_input: unknown) => {
+  loading.value = true;
 
   try {
-    const { content, totalElements } = await props.fetch(
-      pagination.current - 1,
-      pagination.pageSize
-    )
-    data.value = content
-    // data.value = content.map((item: any, index: number) => ({
-    //   ...item
-    // ...Object.keys(item).reduce((acc: any, key) => {
-    //   if (isString(item[key])) acc[key] = truncate(item[key], { length: 30 })
-    //   return acc
-    // }, {}),
-    // row: (pagination.current - 1) * pagination.pageSize + index + 1,
-    // key: item.id
-    // }))
-
-    pagination.total = totalElements
+    const { content, totalElements } = await props.fetch(_input);
+    data.value = content;
+    pagination.total = totalElements;
   } catch (error) {
-    console.error('Error fetching data:', error)
+    console.error("Error fetching data:", error);
   } finally {
-    loading.value = false
+    loading.value = false;
   }
-}
+};
 
 const resetTable = async () => {
-  currentData.value = {
+  await fetchData({
     size: pagination.pageSize,
-    page: pagination.current - 1
-  }
-  await fetchData()
-}
+    page: pagination.current - 1,
+  });
+};
 interface SortParams {
-  sortBy?: string
-  sortDir?: string
+  sortBy?: string;
+  sortDir?: string;
 }
 
-const getSortParams = (sorter: SorterResult<any> | SorterResult<any>[] | null): SortParams => {
-  if (!sorter) return {}
-  const { order, columnKey } = Array.isArray(sorter) ? sorter[0] : sorter
+const getSortParams = (
+  sorter: SorterResult<any> | SorterResult<any>[] | null
+): SortParams => {
+  if (!sorter) return {};
+  const { order, columnKey } = Array.isArray(sorter) ? sorter[0] : sorter;
 
   return {
     sortBy: order ? String(columnKey) : undefined,
     sortDir: order
-      ? order === 'ascend' && columnKey
+      ? order === "ascend" && columnKey
         ? SortDirEnum.ASCENDING
         : SortDirEnum.DESCENDING
-      : undefined
-  }
-}
+      : undefined,
+  };
+};
 
-const updatePagination = (pag: TablePaginationConfig, pagination: PaginationConfig) => {
-  pagination.current = pag.current ?? 1
-  pagination.pageSize = pag.pageSize ?? 5
-}
-const handleTableChange: TableProps['onChange'] = (
+const updatePagination = (
+  pag: TablePaginationConfig,
+  pagination: PaginationConfig
+) => {
+  pagination.current = pag.current ?? 1;
+  pagination.pageSize = pag.pageSize ?? 5;
+};
+const handleTableChange: TableProps["onChange"] = (
   pag: TablePaginationConfig,
   filters: Record<string, FilterValue | null>,
   sorter: SorterResult<any> | SorterResult<any>[],
   extra: TableCurrentDataSource<any>
 ) => {
-  updatePagination(pag, pagination)
-  // console.log('handleTableChange')
+  updatePagination(pag, pagination);
 
-  const { sortBy, sortDir } = getSortParams(sorter)
-  currentData.value = {
+  const { sortBy, sortDir } = getSortParams(sorter);
+  currentData.value = fetchData({
     size: pagination.pageSize,
     page: pagination.current - 1,
     sortBy,
-    sortDir
-  }
-  fetchData()
-}
+    sortDir,
+    ...filters,
+    ...extra,
+  });
+};
 
 const resetState = () => {
-  state.value = undefined
-  state.label = undefined
-}
-const handleTimeSubmit = (selectedKeys: string[], confirm: () => void, dataIndex: string) => {
-  // console.log('🚀 ~ handleTimeSubmit ~ selectedKeys:', selectedKeys)
-  confirm()
+  state.value = undefined;
+  state.label = undefined;
+};
+const handleTimeSubmit = (
+  selectedKeys: string[],
+  confirm: () => void,
+  dataIndex: string
+) => {
+  confirm();
   state.value = selectedKeys[0]
     ? [dayjs(selectedKeys[0][0]).format(), dayjs(selectedKeys[0][1]).format()]
-    : undefined
-  state.label = dataIndex
+    : undefined;
+  state.label = dataIndex;
 
-  currentData.value = {
+  fetchData({
     size: pagination.pageSize,
     page: pagination.current - 1,
-    [state.label]: state.value
-  }
-  fetchData()
-}
+    [state.label]: state.value,
+  });
+};
 
-const handleSubmit = (selectedKeys: string[], confirm: () => void, dataIndex: string) => {
-  // console.log('🚀 ~ handleSubmit ~ selectedKeys:', selectedKeys)
-  state.value = selectedKeys[0]
-  state.label = dataIndex
+const handleSubmit = (
+  selectedKeys: string[],
+  confirm: () => void,
+  dataIndex: string
+) => {
+  state.value = selectedKeys[0];
+  state.label = dataIndex;
 
-  currentData.value = {
+  fetchData({
     size: pagination.pageSize,
     page: pagination.current - 1,
-    [state.label]: state.value
-  }
-  fetchData()
+    [state.label]: state.value,
+  });
   // confirm()
-}
+};
 
 const handleReset = (clearFilters: (arg0: { confirm: boolean }) => void) => {
-  clearFilters({ confirm: true })
-  resetState()
-  currentData.value = {
-    size: pagination.pageSize,
-    page: pagination.current - 1
-  }
-  fetchData()
-}
+  clearFilters({ confirm: true });
+  resetState();
 
-const isDateColumn = (dataIndex: string) => ['createdAt', 'updatedAt'].includes(dataIndex)
+  fetchData({
+    size: pagination.pageSize,
+    page: pagination.current - 1,
+  });
+};
+
+const isDateColumn = (dataIndex: string) =>
+  ["createdAt", "updatedAt"].includes(dataIndex);
 
 onMounted(async () => {
-  currentData.value = {
+  await fetchData({
     size: pagination.pageSize,
-    page: pagination.current - 1
-  }
-  await fetchData()
-})
+    page: pagination.current - 1,
+  });
+});
 
 watch(
   () => props.reloadFlag,
   async () => {
-    await fetchData()
+    await fetchData({
+      size: pagination.pageSize,
+      page: pagination.current - 1,
+    });
   }
-)
+);
 </script>
 
 <style lang="less">